+++ ifen/src/main/java/me/ifen/api/ifen/services/SupportService.java
@@ -9,5 +9,5 @@ public interface SupportService {
 
     public void delete(Long id, Long creatorId);
 
-    public Integer count(Long id);
+    public long count(Long id);
 }
